requirements.txt
streamlit
supabase
pandas
openpyxl
pdfplumber
re
datetime
tempfile
io
xendit-python